fix org-name generation
