+++ packages/fern-dashboard/src/app/actions/inviteUserToOrg.tsx
@@ -1,9 +1,10 @@
 "use server";
 
+import { getAuth0ManagementClient } from "../services/auth0/getAuth0ManagementClient";
+import { getCurrentSession, getOrgMembers } from "../services/auth0/helpers";
+import { Auth0OrgID } from "../services/auth0/types";
+
 /* eslint-disable turbo/no-undeclared-env-vars */
-import { getAuth0ManagementClient, getCurrentSession } from "./auth0";
-import { getOrgMembers } from "./getOrgMembers";
-import { Auth0OrgID } from "./types";
 
 export async function inviteUserToOrg({
   orgId,
