+++ packages/fern-dashboard/src/app/services/venus/helpers.ts
@@ -1,21 +1,6 @@
-/* eslint-disable turbo/no-undeclared-env-vars */
-import { FernVenusApi, FernVenusApiClient } from "@fern-api/venus-api-sdk";
+"use server";
 
-export function getVenusClient({
-  token,
-}: {
-  token: string;
-}): FernVenusApiClient {
-  if (process.env.VENUS_SERVER_URL == null) {
-    throw new Error(
-      "VENUS_SERVER_URL is not defined in the current environment"
-    );
-  }
-  return new FernVenusApiClient({
-    environment: process.env.VENUS_SERVER_URL,
-    token,
-  });
-}
+import { FernVenusApi, FernVenusApiClient } from "@fern-api/venus-api-sdk";
 
 export async function isMemberOf(
   venus: FernVenusApiClient,
