+++ packages/fern-dashboard/src/components/layout/AppLayout.tsx
@@ -17,7 +17,7 @@ export async function AppLayout({ children }: AppLayout.Props) {
     <div className="dark:bg-gray-1200 flex flex-1 flex-col bg-gray-300">
       <Header session={session} />
       <div className="flex min-h-0 flex-1 flex-col md:flex-row-reverse">
-        <div className="dark:border-gray-1100 flex flex-1 justify-center overflow-y-auto border border-gray-500 bg-white px-12 pt-12 md:mr-4 md:rounded-t-2xl dark:bg-black">
+        <div className="dark:border-gray-1100 flex flex-1 justify-center overflow-y-auto border-x border-t border-gray-500 bg-white px-12 pt-12 md:mr-4 md:rounded-t-2xl dark:bg-black">
           <div className="flex max-w-[1200px] flex-1 flex-col">
             <div className="flex flex-1">{children}</div>
             <div className="py-12">
